Department of State: Languages
Moved view code to separate file.

#2880
